requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.4
python-dotenv==1.0.1
scikit-learn==1.4.1.post1
xgboost==2.0.3
pandas==2.2.1
numpy==1.26.4
httpx==0.27.2
pytest==8.1.1
httpx==0.27.2
evident==0.4.30
